Diseño e Implementación de un Sistema de Clasificación Contable Automatizado para Facturas/src/Funciones Cloud/Clasificador Automático/requirements.txt
fuzzywuzzy[python-Levenshtein]
google-cloud-firestore
google-cloud-storage
scikit-learn
Flask==2.3.3
voluptuous
joblib
numpy
nltk==3.8.1